requirements.txt
Pillow
matplotlib
numpy
inference-sdk
